Branch jmp_scr4_cleanup
Minor changes to (a lot of) scr4 source files to avoid warnings:
    - (size_t) cast to (int) in calls to strlen, sizeof...
    - unused local variables removed from declarations
    - system calls replaced by new MSVC standard (unlink -> _unlink...)
    - (f)printf parameters in excess removed
    - some function headers reformated in C++ std
    - returned value specified for functions returning an int (std C++)
    - missing (not declared) functions added in .h files in the C++ std

+++ scr4/s_dirstd.c
@@ -26,7 +26,7 @@ char    *res, *dir;
 #if defined(DOS) && !defined(DOSW32) && !defined(SCRW32)
     SCR_lower(res);
 #endif
-    lg = strlen(res);
+    lg = (int)strlen(res);
     if(lg != 1 && U_is_in(res[lg - 1], "/\\") &&    /* \ */
        !(lg == 3 && res[1] == ':'))                 /* a:\ */
 	res[lg - 1] = 0;
