+++ scr4/s_ststr.c
@@ -13,7 +13,7 @@ int     len;
     short   i;
     TBLSTR  *tb;
 
-    if(len == 0) len = strlen(txt) + 1;
+    if(len == 0) len = (int)strlen(txt) + 1;
     i = SCR_str_find_str(txt, len);
     if(i >= 0) return(i);
     if(NB_TBL_STR == 0)
@@ -39,7 +39,7 @@ int     len;
     short   i;
     TBLSTR  *tb;
 
-    if(len == 0) len = strlen(txt) + 1;
+    if(len == 0) len = (int)strlen(txt) + 1;
     for(i = 0 , tb = TBL_STR ; i < NB_TBL_STR ; i++, tb++)
 	if(memcmp(txt, tb->ts_ptr, min(len, tb->ts_len)) == 0)
 	    return(i);
